backend and frontend adjusted for both regged users and not regged users

+++ backend/index.js
@@ -177,8 +177,28 @@ const getUncryptedCartId = (cryptedId) => {
     return cartIdFromDB;
 }
 
-const handleCartIdBeforeDB = (reqCartId) => {
-    const cartIdFromFrontend = decodeURIComponent(reqCartId);
+const handleCartIdBeforeDB = (reqBody, reqParams) => {
+    let bodyOrParams;
+    for (let i in reqBody) {
+        if (i === 'cartId') {
+            bodyOrParams = 'body';
+        }
+    }
+    if (bodyOrParams !== 'body') {
+        for (let i in reqParams) {
+            if (i === 'cartid') {
+                bodyOrParams = 'params';
+            }
+        }
+    }
+
+    let cartIdFromFrontend;
+    if (bodyOrParams === 'body') {
+        cartIdFromFrontend = reqBody.cartId;
+    } else {
+        cartIdFromFrontend = decodeURIComponent(reqParams.cartid);
+    }
+
     let plainCartId;
     const regexOnlyNumbers = new RegExp('^[0-9]+$', 'gi');
     if (regexOnlyNumbers.test(cartIdFromFrontend)) {
@@ -186,10 +206,11 @@ const handleCartIdBeforeDB = (reqCartId) => {
     } else {
         plainCartId = getUncryptedCartId(cartIdFromFrontend);
     }
+
     return plainCartId;
 }
 
-// Create a new cart
+// Create a new logged out cart
 app.post('/api/cart/new', (req, res) => {
     let stmt = dbWatches.prepare(`
         INSERT INTO carts
@@ -222,7 +243,7 @@ app.post('/api/cart/new', (req, res) => {
 // Add a product to the cart - or - add more of the same product (quantity)
 app.post('/api/cart/add/productid/:productid', (req, res) => {
     // First get the uncrypted cartId that belongs to the users crypted one
-    const cartIdFromDB = getUncryptedCartId(req.body.cartId);
+    const plainCartId = handleCartIdBeforeDB(req.body, req.params);
 
     // Then use the uncrypted one to make changes in DB
     let stmt = dbWatches.prepare(`
@@ -230,7 +251,7 @@ app.post('/api/cart/add/productid/:productid', (req, res) => {
         VALUES (:cartIdParam, :productIdParam)
     `);
     let info = stmt.run({
-        cartIdParam: cartIdFromDB,
+        cartIdParam: plainCartId,
         productIdParam: req.params.productid
     });
     res.json({'Additions made': info.changes})
@@ -238,7 +259,7 @@ app.post('/api/cart/add/productid/:productid', (req, res) => {
 
 // Get all items in your cart
 app.get('/api/cart/:cartid', (req, res) => {
-    let plainCartId = handleCartIdBeforeDB(req.params.cartid);
+    let plainCartId = handleCartIdBeforeDB(req.body, req.params);
 
     let stmt = dbWatches.prepare(`
         SELECT products.*,
@@ -255,8 +276,7 @@ app.get('/api/cart/:cartid', (req, res) => {
 
 // Delete a product from cart, no matter the quantity of that product.
 app.delete('/api/cart/:cartid/removefromcart/productid/:productid', (req, res) => {
-    const cryptedCartId = decodeURIComponent(req.params.cartid);
-    const cartIdFromDB = getUncryptedCartId(cryptedCartId);
+    let plainCartId = handleCartIdBeforeDB(req.body, req.params);
 
     let stmt = dbWatches.prepare(`
         DELETE FROM cartItems
@@ -264,7 +284,7 @@ app.delete('/api/cart/:cartid/removefromcart/productid/:productid', (req, res) =
         AND productId = :productIdParam
     `);
     let result = stmt.run({
-        cartIdParam: cartIdFromDB,
+        cartIdParam: plainCartId,
         productIdParam: req.params.productid
     })
     res.json(result);
@@ -272,8 +292,7 @@ app.delete('/api/cart/:cartid/removefromcart/productid/:productid', (req, res) =
 
 // Delete one, "Quantity - 1"
 app.delete('/api/cart/:cartid/quantitydecrease/productid/:productid', (req, res) => {
-    const cryptedCartId = decodeURIComponent(req.params.cartid);
-    const cartIdFromDB = getUncryptedCartId(cryptedCartId);
+    let plainCartId = handleCartIdBeforeDB(req.body, req.params);
 
     let stmt = dbWatches.prepare(`
         DELETE FROM cartItems
@@ -285,7 +304,7 @@ app.delete('/api/cart/:cartid/quantitydecrease/productid/:productid', (req, res)
         )
     `);
     let result = stmt.run({
-        cartIdParam: cartIdFromDB,
+        cartIdParam: plainCartId,
         productIdParam: req.params.productid
     })
     res.json(result);
@@ -309,25 +328,33 @@ app.patch('/api/cart/movetopersonalcart', (req, res) => {
     res.json(result);
 })
 
-// When purchase is done, delete cart and all products in cart from DB
-app.delete('/api/cart/:cartid/removecartcompletely', (req, res) => {
-    const cryptedCartId = decodeURIComponent(req.params.cartid);
-    const cartIdFromDB = getUncryptedCartId(cryptedCartId);
+// When purchase is done, for not logged in users: 
+// delete cart and all products in cart from DB
+app.delete('/api/cart/:cartid/cartcheckoutdb', (req, res) => {
+    let plainCartId = handleCartIdBeforeDB(req.body, req.params);
+    const cartIdParams = decodeURIComponent(req.params.cartid);
+    const regexOnlyNumbers = new RegExp('^[0-9]+$', 'gi');
 
     let stmt = dbWatches.prepare(`
         DELETE FROM cartItems
         WHERE cartId = :cartIdParam
     `);
     let result = stmt.run({
-        cartIdParam: cartIdFromDB
+        cartIdParam: plainCartId
     })
-    let stmt2 = dbWatches.prepare(`
-        DELETE FROM carts
-        WHERE cartId = :cartIdParam
-    `);
-    let result2 = stmt2.run({
-        cartIdParam: cartIdFromDB
-    })
+
+    let stmt2, result2;
+    if (!regexOnlyNumbers.test(cartIdParams)) {
+        // oinloggad varukorg, deleta även själva varukorgen
+        stmt2 = dbWatches.prepare(`
+            DELETE FROM carts
+            WHERE cartId = :cartIdParam
+        `);
+        result2 = stmt2.run({
+            cartIdParam: plainCartId
+        })
+    }
+    
     res.json({"cartItems": result, "carts": result2});
 })
 //=================
